Reorganising pages folders
