+++ src/pages/Signup/index.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useContext } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import authService from "../services/auth.service";
-import loginStyles from "./styles/Login.module.sass";
+import authService from "../../services/auth.service";
+import loginStyles from "../Login/index.module.sass";
 
 function SignupPage() {
   const [email, setEmail] = useState("");
